+++ src/AttributesBlockParser.php
@@ -11,7 +11,6 @@
 
 namespace Webuni\CommonMark\AttributesExtension;
 
-use League\CommonMark\Block\Element\Document;
 use League\CommonMark\Block\Parser\AbstractBlockParser;
 use League\CommonMark\ContextInterface;
 use League\CommonMark\Cursor;
@@ -20,19 +19,6 @@ class AttributesBlockParser extends AbstractBlockParser
 {
     public function parse(ContextInterface $context, Cursor $cursor)
     {
-        $document = $context->getDocument();
-        $tip = $context->getTip();
-
-        if (!$document->lastChild() instanceof AttributesDocument) {
-            $attributesDocument = new AttributesDocument();
-            $attributesDocument->replaceChildren($document->children());
-            $document->replaceChildren([$attributesDocument]);
-
-            if ($tip instanceof Document) {
-                $context->setTip($attributesDocument);
-            }
-        }
-
         $state = $cursor->saveState();
         $attributes = AttributesUtils::parse($cursor);
         if (empty($attributes)) {