+++ CHANGELOG.md
@@ -1,7 +1,24 @@
 CHANGELOG
 =========
 
-0.1.0
------
+0.4.0 (2016-07-16)
+------------------
+
+ * Refactor code to use CommonMark 0.14 API
+ * Fix non trimmed spaces for inline attributes related to blocks
+ * Fix order of attributes
+
+0.3.0 (2015-09-26)
+------------------
+
+ * Refactor code to use CommonMark 0.11 API 
+
+0.2.0 (2015-07-27)
+------------------
+
+ * Use new CommonMark 0.10 API for element attributes
+
+0.1.0 (2015-07-13)
+------------------
 
  * Initial implementation